+++ app/assets/javascripts/application.js
@@ -21,3 +21,6 @@
 //= require external/lang-matlab
 //= require external/lang-r
 //= require user_voice
+
+//= require external/highcharts-4.0.1.src.js
+//= require variables